feat(dispute): 관리자 분쟁 해결 API 구현 완료
- DisputeService: ResolveDisputeAsync, GetAllDisputesAsync, GetDisputeDetailForAdminAsync 구현
- DisputeRepository: GetAllDisputesCursorAsync, GetDisputeByIdAsync 구현
- 토스 환불 연동 (RESOLVED_BUYER 시 CancelPaymentAsync 호출)
- DISPUTE_RESOLVED 알림 구매자/판매자 양쪽 발송
- EF 재스캐폴드 사이드 이펙트 수정 (int→long 타입 변경)
- AdminDisputeResolutionTests: 10개 E2E 테스트 GREEN
- DB 덤프 업데이트

+++ TicketPlatFormServer/Repository/Sell/ISellRepository.cs
@@ -83,7 +83,7 @@ public interface ISellRepository
     /// <summary>
     /// 특정 티켓의 이미지 목록 조회
     /// </summary>
-    Task<List<TicketImage>> GetTicketImagesByTicketIdAsync(int ticketId);
+    Task<List<TicketImage>> GetTicketImagesByTicketIdAsync(long ticketId);
 
     /// <summary>
     /// 사용자의 판매 티켓 목록 조회 (페이징)
