+++ TicketPlatFormServer/Repository/Transaction/ITransactionRepository.cs
@@ -68,5 +68,5 @@ public interface ITransactionRepository
     /// <returns>TransactionStatus (없으면 null)</returns>
     Task<DBModel.TransactionStatus?> GetTransactionStatusByCodeAsync(string code);
 
-    Task<PaymentPreviewReadModel?> GetPaymentPreviewAsync(long transactionId, int buyerId);
+    Task<PaymentPreviewReadModel?> GetPaymentPreviewAsync(long transactionId, long buyerId);
 }
